requirements.txt
browser-use>=0.1.0
python-dotenv>=1.0.1
requests>=2.32.3
beautifulsoup4>=4.12.3
openai>=1.43.0
numpy>=2.1.3
pandas>=2.2.3
scikit-learn>=1.5.2
matplotlib>=3.9.2
